지도 수정 (사용자 컨트롤 추가) js 파일 분리
