+++ src/components/PostDetailMap.jsx
@@ -1,58 +1,19 @@
 /*global kakao*/
 import { useEffect } from 'react';
+import { getPostLocation } from '../api/getmap';
 
 const PostDetailMap = ({ post }) => {
   useEffect(() => {
-    // 스크립트 로드가 완료된 후 실행
-    const container = document.getElementById('map'); //지도를 담을 영역의 DOM 레퍼런스
-    const options = {
-      //지도를 생성할 때 필요한 기본 옵션
-      center: new kakao.maps.LatLng(33.450701, 126.570667), //지도의 중심좌표.
-      level: 3 //지도의 레벨(확대, 축소 정도)
-    };
-    var map = new kakao.maps.Map(container, options); //지도 생성 및 객체 리턴
-    console.log('map===>', map);
-
     var services = kakao.maps.services;
     if (!services) {
       console.log('kakao.maps.services 를 찾을 수 없음 :', kakao.maps);
       return;
     } else {
-      // 주소-좌표 변환 객체를 생성합니다
-      var geocoder = new kakao.maps.services.Geocoder();
-
-      // 가게주소
-      const address = post.address;
-      // 주소로 좌표를 검색합니다
-      geocoder.addressSearch(address, function (result, status) {
-        // 정상적으로 검색이 완료됐으면
-        if (status === kakao.maps.services.Status.OK) {
-          var coords = new kakao.maps.LatLng(result[0].y, result[0].x);
-
-          // 결과값으로 받은 위치를 마커로 표시합니다
-          var marker = new kakao.maps.Marker({
-            map: map,
-            position: coords
-          });
-
-          // 인포윈도우로 장소에 대한 설명을 표시합니다
-          var infowindow = new kakao.maps.InfoWindow({
-            content: `<div style="width:150px;text-align:center;padding:6px 0;">${post.title}</div>`
-          });
-          infowindow.open(map, marker);
-
-          // 지도의 중심을 결과값으로 받은 위치로 이동시킵니다
-          map.setCenter(coords);
-        }
-      });
+      getPostLocation(post);
     }
   }, []);
 
-  return (
-    <div id="map" className="w-[500px] h-[400px] bg-slate-50">
-      지도담을영역
-    </div>
-  );
+  return <div id="map" className="w-[500px] h-[400px] bg-slate-50"></div>;
 };
 
 export default PostDetailMap;
